Improve the Chain of Thoughts experiment to allow successive planning and resolution phases in the same conversation.

+++ chain_of_thoughts.py
@@ -33,13 +33,13 @@ LLMs = [
 DEFAULT_INSTRUCTOR = 'mistral-large'
 DEFAULT_EXECUTOR = DEFAULT_INSTRUCTOR
 
-SYSTEM_COT = 'Break down the complex problem from the prompt and derive a very detailed plan to solve the problem. Write the plan as a flat list of tasks.\n'
+SYSTEM_COT = 'Break down the complex problem into smaller pieces and create a very detailed plan to solve each piece separately. Be careful to include everything required to solve the problem and to avoid unecessary work. Specify each task in natural language with as many details as possible. Write the plan as a long and flat list of tasks without any hierarchy.\n'
 
-SYSTEM_PLAN = 'Extract all tasks and output them as a JSON list of strings.\n'
+SYSTEM_PLAN = 'Extract all tasks with their full description and output them as a JSON list of strings.\n'
 
-SYSTEM_TASK = 'Leverage the context to process the task and to progress toward the solution to the user\'s problem without repeating what is already in the context.\n'
+SYSTEM_TASK = 'Leverage the whole context to process the task without repeating what is already in the context.\n'
 
-SYSTEM_PROMPT = 'Leverage the context to answer the prompt.'
+SYSTEM_PROMPT = 'Leverage the whole context to answer the prompt.'
 
 STATE_FILE = 'cot_session_state.json'
 
@@ -92,7 +92,8 @@ st.session_state.do_planning = st.checkbox('Plan', value=st.session_state.do_pla
 
 
 if st.button('Reset', type='primary', disabled=not os.path.exists(STATE_FILE)):
-    os.unlink(STATE_FILE)
+    if os.path.exists(STATE_FILE):
+        os.unlink(STATE_FILE)
     state = {
         'instructor': DEFAULT_INSTRUCTOR,
         'executor': DEFAULT_EXECUTOR,
@@ -114,6 +115,17 @@ if len(st.session_state.history) > 0:
             st.markdown(item['message'])
 
 
+def build_prompt(problem=None):
+    if problem:
+        prompt = f"User problem:\n{problem}\n\n\n"
+    else:
+        prompt = ''
+    prompt += 'Context:\n\n'
+    for item in st.session_state.history:
+        prompt += f"\t{item['role']}: {item['message']}\n\n"
+    return prompt
+
+
 if st.session_state.prompt is None:
     if prompt := st.chat_input('Submit a complex problem...'):
         st.session_state.prompt = prompt
@@ -126,10 +138,8 @@ elif st.session_state.do_planning and st.session_state.plan is None:
     with st.chat_message('user'):
         st.markdown(prompt)
 
-    prompt2 = 'Context:\n\n'
-    for item in st.session_state.history:
-        prompt2 += f"\t{item['role']}: {item['message']}\n\n"
-    prompt2 += '## Prompt:\n' + prompt + '\n\n\n' + SYSTEM_COT
+    prompt2 = build_prompt()
+    prompt2 += '## Problem:\n' + prompt + '\n\n\n' + SYSTEM_COT
     response = llm(prompt2, INSTRUCTOR)
     with st.chat_message('ai'):
         answer = st.write_stream(streaming_callback(response))
@@ -143,8 +153,6 @@ elif st.session_state.do_planning and st.session_state.plan is None:
 
 elif st.session_state.do_planning and st.session_state.go:
     st.session_state.history.append({'role': 'ai', 'message': st.session_state.plan})
-    with st.chat_message('assistant'):
-        st.markdown(st.session_state.plan)
 
     tstart = datetime.now()
     
@@ -170,9 +178,7 @@ elif st.session_state.do_planning and st.session_state.go:
         with st.chat_message('assistant'):
             st.markdown('### ' + step)
         
-        prompt2 = 'Context:\n\n'
-        for item in st.session_state.history:
-            prompt2 += f"\t{item['role']}: {item['message']}\n\n"
+        prompt2 = build_prompt(st.session_state.prompt)
         prompt2 += f"\n\n\n## Task\n{step}"
         prompt2 += '\n\n\n' + SYSTEM_TASK 
         response = llm(prompt2, EXECUTOR)
@@ -208,14 +214,14 @@ elif st.session_state.do_planning and st.session_state.go:
     st.rerun()
 
 else:
+    st.chat_input('Busy...', disabled=True)
+    
     prompt = st.session_state.prompt
     with st.chat_message('user'):
         st.markdown(prompt)
     st.session_state.history.append({'role': 'user', 'message': prompt})
     
-    prompt2 = 'Context:\n\n'
-    for item in st.session_state.history:
-        prompt2 += f"\t{item['role']}: {item['message']}\n\n"
+    prompt2 = build_prompt()
     prompt2 += f"\n\n\n## Task\n{prompt}"
     prompt2 += '\n\n\n' + SYSTEM_PROMPT
     response = llm(prompt2, INSTRUCTOR)
